Diff editor code cleanup (#199761)
* Diff editor code cleanup
