+++ src/vs/editor/browser/widget/diffEditor/components/accessibleDiffViewer.css
