+++ src/vs/editor/browser/widget/diffEditor/components/diffEditorSash.ts
@@ -6,7 +6,7 @@
 import { IBoundarySashes, ISashEvent, Orientation, Sash, SashState } from 'vs/base/browser/ui/sash/sash';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { IObservable, IReader, autorun, derived, observableValue } from 'vs/base/common/observable';
-import { DiffEditorOptions } from './diffEditorOptions';
+import { DiffEditorOptions } from '../diffEditorOptions';
 
 export class DiffEditorSash extends Disposable {
 	private readonly _sashRatio = observableValue<number | undefined>(this, undefined);
