+++ src/vs/editor/browser/widget/diffEditor/components/diffEditorViewZones/renderLines.ts
