+++ src/vs/editor/browser/widget/diffEditor/diffEditor.contribution.ts
@@ -18,6 +18,7 @@ import { CommandsRegistry } from 'vs/platform/commands/common/commands';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { ContextKeyEqualsExpr, ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
+import './registrations.contribution';
 
 export class ToggleCollapseUnchangedRegions extends Action2 {
 	constructor() {
