+++ src/vs/editor/browser/widget/diffEditor/features/outlineModel.ts
